Extract impl of pixel feeder
